requirements.txt
# Backend
fastapi>=0.111.0
uvicorn>=0.29.0
openai>=1.30.0
pydantic>=2.7.0
python-dotenv>=1.0.1

# Métricas
torch>=2.3.0
transformers>=4.41.0
nltk>=3.8.1

# Frontend
streamlit>=1.35.0
requests>=2.31.0
